API/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docker==4.2.2
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
protobuf==3.12.2
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
